refactor(datagrid): sort icon component
